+++ R/LandR-package.R
@@ -31,7 +31,7 @@
 #' @importFrom ggplot2 scale_fill_distiller scale_fill_manual
 #' @importFrom ggplot2 scale_fill_viridis_c scale_fill_viridis_d
 #' @importFrom ggplot2 scale_linetype_manual scale_x_discrete
-#' @importFrom ggplot2 stat stat_summary sym theme theme_classic unit
+#' @importFrom ggplot2 stat stat_summary sym theme theme_bw theme_classic unit
 #' @importFrom ggpubr theme_pubr
 #' @importFrom ggspatial annotation_north_arrow layer_spatial north_arrow_minimal
 #' @importFrom grDevices colorRampPalette dev.off png
@@ -68,5 +68,6 @@
 #' @importFrom tools file_ext file_path_sans_ext
 #' @importFrom utils capture.output combn count.fields data getFromNamespace head install.packages
 #' @importFrom utils str tail untar
+#' @importFrom viridis scale_fill_viridis
 ## usethis namespace: end
 NULL